* develop: (21 commits)
  NF: Cleaned up environment vars & grunt tasks
  NF: Only include mix panel in prod
  BF: Try running npm install with --force flag
  RF: Moved all vendor files into separate sub-dir in config to centralize.  I FEEL SO MUCH BETTER NOW
  RF: Massive cleanup of console-specific env files.  CSS Libs were all fucked up, minified them all
  NF: Create secure server in dev as well
  BF: change to explicitly require async 0.9.2
  BF: Add repository field and install grunt-cli globally
  BF: Update grunt node inspector
  NF: Add mocha to dev dependencies
  BF: Grunt back into dev dependencies...
  NF: Run https_redirect in deploy, make grunt explicit dependency
  BF
  NF: add local grunt
  NF: Upgrade gcloud to 0.27.0
  NF: Switch to develop utils
  NF: update deploy_console
  NF Add config pull to activate script
  NF: Activate env script replaces activate_production
  RF: Change name of "development" env to "dev" to be consistent with other repos
  ...

Conflicts:
	config/env/dev.js
	config/express.js

+++ deploy_console.sh
@@ -1,28 +1,84 @@
 #!/bin/bash
 
-source ./activate_production.sh
-nvm use 0.12.0
-npm install
-rm -rf ./public/dist/*
-grunt build
+# usage text visible when --help flag passed in
+usage="$(basename "$0") -- deploy the Cliques Console
 
-if [ ! -d $HOME"/repositories/cliques-config" ]; then
-    git clone git@github.com:cliquesads/cliques-config.git ../cliques-config
-    ln -s ../cliques-config .
-else
-    cd ../cliques-config
-    git pull
-    cd ../cliques-console
+where:
+    --help  show this help text
+    -e arg (='production') environment flag - either 'dev' or 'production'.  Defaults to production"
+
+# BEGIN environment parsing
+env="production"
+
+if [ ! -z $1 ]; then
+  if [ $1 == '--help' ]; then
+    echo "$usage"
+    exit 0
+  fi
 fi
 
-processname='cliques-console'
+# fucking getopts
+while getopts ":e:" opt; do
+  case $opt in
+    e)
+      if [ "$OPTARG" != 'production' ] && [ "$OPTARG" != 'dev' ]; then
+        echo "Invalid environment: $OPTARG.  Environment must be either 'dev' or 'production'"
+        exit 1
+      else
+        env="$OPTARG"
+      fi
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      echo "$usage"
+      exit 1
+      ;;
+    :)
+      echo "Environment flag -$OPTARG requires an argument (either 'dev' or 'production')" >&2
+      exit 1
+      ;;
+  esac
+done
+# END environment parsing
+
+############### BEGIN Environment setup ################
+
+# Set proper environment variables now that env is set
+if [ "$env" == "production" ]; then
+    processname='cliques-console'
+else
+    processname='cliques-console-dev'
+fi
+
+source activate_env.sh -e $env
+# if activate_env failed then bail
+if [ $? -ne 0 ]; then
+    exit $?
+fi
+
+# run npm install to install any new dependencies
+npm install --force
+npm install cliques-node-utils
+
+# run build step
+rm -rf ./public/dist/*
+grunt build-"$env"
+
 running=$(pm2 list -m | grep "$processname")
+redir_running=$(pm2 list -m | grep "https_redirect")
 
 if [ -z "$running" ]; then
-#    # hook PM2 up to web monitoring with KeyMetrics
-    pm2 link pmmeigey04ri9gk qzqytpy0n48vh2z  $hostname
+    # hook PM2 up to web monitoring with KeyMetrics
+    pm2 link $KEYMETRICS_PRIVATE_KEY $KEYMETRICS_PUBLIC_KEY $HOSTNAME
     # start in cluster mode
     pm2 start server.js --name "$processname" -i 0
 else
     pm2 gracefulReload "$processname"
+fi
+
+if [ -z "$redir_running" ]; then
+    # start in cluster mode
+    pm2 start https_redirect.js -i 1 --name "https_redirect"
+else
+    pm2 gracefulReload "https_redirect"
 fi
